Experiment to make slicer work with binary python windows download
This version is hardcoded for testing, but shows (some of?) what needs
to change to support this method.

To build the binary python directory, the core python was downloaded
from python.org and was augmented by a copy of Itcl3.4 downloaded as part
of the activetcl package (both python and activetcl used tcl8.5 in
x64 release mode so the binaries are compatibile).  Then an x64 build
of numpy was added to the directory using a download from a site pointed
to from scipy.org.