BUG: Call AddNode when restoring a MRML scene view
Scene events must always be fired, observers can decide to ignore them or not.
Some logics can't afford to ignore events and must be always in-sync with the scene, they should receive all the events, even when restoring a scene view.

vtkMRMLSceneViewNodeEventsTest was failing because NodeAboutToBeAddedEvent was never fired (but NodeAddedEvent was) when restoring a scene.
It used to be AddNode, but r15757 changed it to AddNodeNoNotify to prevent a crash after a scene is restored.
The crash doesn't seem to be there anymore.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@18879 3bd1e089-480b-0410-8dfb-8563597acbee