COMP: Change stdio.h, errno.h, stdlib.h into their corresponding c++ includes.
See http://www.cplusplus.com/reference/clibrary/

The C++ library includes the same definitions as the C language library
organized in the same structure of header files, with the following
differences:

    Each header file has the same name as the C language version but
with a "c" prefix and no extension. For example, the C++ equivalent
for the C language header file <stdlib.h> is <cstdlib>.
    Every element of the library is defined within the std namespace.

Nevertheless, for compatibility with C, the traditional header names
name.h (like stdlib.h) are also provided with the same definitions
within the global namespace. In the examples provided in this reference,
this version is used so that the examples are fully C-compatible,
although its use is deprecated in C++.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@18967 3bd1e089-480b-0410-8dfb-8563597acbee