BUG: Don't update current index before QComboBox::currentIndexChanged(QString) is emitted
There is a bug in Qt that makes doing actions dangerous when QComboBox::currentIndex(int) is fired :
void QComboBoxPrivate::_q_emitCurrentIndexChanged(const QModelIndex &index)
{
    Q_Q(QComboBox);
    emit q->currentIndexChanged(index.row());
    emit q->currentIndexChanged(itemText(index));
}
if the QComboBox current index is changed when the SIGNAL currentIndexChanged(int) is fired, the QModelIndex index is then no longer valid when emit q->currentIndexChanged(itemText(index)); is reached.
This is what happens when the color table combobox model (qMRMLColorTableComboBox) is initialized (qMRMLSceneModel::populateFromScene()), the 1 item added into the model is a category which is automatically selected (because the combobox was empty before), which emits the currentIndexChanged(int) signal that was being caught by qMRMLNodeComboBox that was automatically unselecting the index (because category is not a valid entry for the color table combobox, only color nodes are selectable). when the second signal currentIndexChange(QString) is fired, the current index was already changed and index (in _q_emitCurrentIndexChanged) was not valid (garbage) anymore.

As a (no better) solution, qMRMLNodeComboBox now does actions on the currentIndexChanged(QString) signal.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@18968 3bd1e089-480b-0410-8dfb-8563597acbee