ENH: Pass CMAKE_DISABLE_FIND_PACKAGE_Java to inner build
On Mac OS X, java is not installed by default.
However, there is a java binary in /usr/bin, when run, the user is asked if he/she wants to install Java.
So for CMake, Java is found (which is true). There need a way for the user to prevent Java to be found.
Solution works only on CMake > 2.8.6

git-svn-id: http://svn.slicer.org/Slicer4/trunk@18988 3bd1e089-480b-0410-8dfb-8563597acbee