ENH: initial checkin of new DICOM plugin interface
Example plugins for Scalar and Diffusion volumes provided.  New interface
will allow separation of DICOM module from the details of different
volume parsing and loading.

Next steps are to flesh out different DICOM datatypes and also to 
support exporting of slicer data to DICOM.  In addition, effiency of the
DICOM header examination process needs to be addressed (to avoid
duplicated parsing in different plugins).  Some of this will need to
happen at the ctkDICOM level.



git-svn-id: http://svn.slicer.org/Slicer4/trunk@19217 3bd1e089-480b-0410-8dfb-8563597acbee