BUG: Activating ViewControllers module would change SliceView content
When linking was enabled on SliceComposite nodes, activating the
ViewControllers module couse the SliceViews to change their content.

ViewControllers module will create new ViewControllers to pack in the
module panel. These controllers need to be configured from MRML. The
UpdateWidgetFromMRML* methods would cause the selection of foreground
and background volumes to propagate through the link logic.

Disabled the signals on the foreground, background, and label map
comboboxes when setting their values inside of UpdateWidgetFromMRML*
methods.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@19320 3bd1e089-480b-0410-8dfb-8563597acbee