ENH: Avoid "flickering" when scene view dialog is opened to edit node
The dialog was first opened as if it is "taking a scene view" then it
it set to edit the node. It leaves enough time to see the change of state
by the user.
Fix is to setup the panel before it is shown.

Side fix: make the dialog modeless (block the user interactions until the
dialog is accepted/rejected). There is no reason to keep it on (or it needs
more feature.)
rejected() and accepted() already hide() the dialog automatically, no need
to manually do it.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@19481 3bd1e089-480b-0410-8dfb-8563597acbee