BUG: restrict morphology operations to scope of 'All'
Really no convenient way to have the operation apply only
to the visible region since it does not get applied
with respect to a slice.

Real solution for these would be to give a scope option of
'Window' in which case the EffectOptions would present a
combo box of the current views and the user could select
the one to apply the operation in - add this to the TODO list.