ENH: adopt explicit declaration of connections for guis
By adding the connections list to each class instance it is
possible for users of the class to explicitly turn off the connection
before updating the widget values.  This avoids having
updateMRMLFromGUI calls interfering with the updating
of the gui itself.

Label and Paint have been updated with this behavior, other
effects to follow.