COMP: On Windows 64b, pass INCR_TCL_VERSION_DOT to inner build
On Windows, tcl is not built but pre-built. Itcl is packaged with tcl and doesn't require the external project itcl. Make sure
INCR_TCL_VERSION_DOT and INCR_TCL_VERSION are passed to the inner build so that their path are added into the launcher configuration file.
Failing test: qSlicerApplicationTpyclTest1
Note that the fix is not enough, and itcl34.dll is still failing to load. It seems to use a bad version of msvcr90.dll

git-svn-id: http://svn.slicer.org/Slicer4/trunk@19570 3bd1e089-480b-0410-8dfb-8563597acbee