COMP: Ensure environment is set when running "out-of-source" scripted modules tests.
This is a follow up to r19827 [1] where it was mentioned that

   "PYTHONPATH is updated so that python module associated with
   a module could be imported. For example, the scripted module."

It resumes that at the time of that commit, this was an incorrect statement.
Indeed, the python macro defined in SlicerMacroPythonTesting where
relying on "Slicer_LAUNCHER_EXECUTABLE" whereas the variable updated
with '--launcher-additional-settings' was "Slicer_LAUNCH_COMMAND".

This commit set the command line option specific to launcher additional settings
into a variable that could be re-used:

    Slicer_ADDITIONAL_LAUNCHER_SETTINGS

The macro SLICER_ADD_PYTHON_TEST and SLICER_ADD_PYTHON_UNITTEST have been
updated to consider that variable.

[1] http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=19827

git-svn-id: http://svn.slicer.org/Slicer4/trunk@19853 3bd1e089-480b-0410-8dfb-8563597acbee