COMP: build DCMTK as external project and use it to build CTK
This will allow us to use the same DCMTK for both CTK and ITKv4.

This move is also motivated by the need to use a version of DCTMK that
includes dcmrt, as needed for the SlicerRT effort.

Currently this build points to the master branch of the dcmkt.git.rt
repository, but should be migrated to a release branch as soon
as one is available.