ENH: use the app rather than the module property for python access to database
These changes remove the dependence on the slicer.dicomDatabase
attribute and instead query the app for the database.  This would be
preferred, but creation of the ctk.ctkDICOMDatabase in python
and passing it to C++ leads to a crash; apparently the python-created
cannot be overwritten with None without leading to a crash.