ENH: Add 'reload-and-test' infrastructure to scripted module template
This provides a one-click method for developers to modify their
code and their test on the fly.  The sequence should be that developers
can create a module (extension) using the ModuleWizard, point
their module path to it, and then start slicer to load
their new module.  From there they just edit the implementation (GUI
and Logic) and the test to exercise it.  Hitting the reload and test
button brings in the latest copy of their code, replaces the GUI, and
runs the test.