PERF: Don't render while the scene is busy
Note: this requires a change to CTK to make forceRender virtual.

Also this don't work for the slice views, which still render during
scene load because there is no qMRMLSliceView (yet) that can
override this method.  Instead, there is the ctkVTKSliceView, which
doesn't know about MRML.  So we need a qMRMLSliceView that inherits from
ctkVTKSliceView and implements this override.