BUG: #2629.  Window level controls conflicted with crosshair navigation
CrosshairDisplayableManager now observe events directly on the Interactor
so the InteractorStyle cannot swallow the events prior to the DisplayableManager
seeing the event.

Added facility in AbstractDisplayableManager to manage events on Interactor
in a fashion similar to how events are managed on the InteractorStyle.

Added priorities to events for Interactor and InteractorStyle

Added Abort flag to events on Interactor.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@21160 3bd1e089-480b-0410-8dfb-8563597acbee