BUG: new test to baseline for bug #1926
Test case for the interaction between the ViewControllers module, linking, slice
interpolation, and the selection of background, foreground, and label images.
When entering the ViewControllers module, extra View Controllers are created and
configured.  If linking is on, then foreground, backgroud, and label selection
can be propagated to the other views incorrectly. If the node selectors are
bocked from emitting signals, then the viewers maintain their proper volumes.
However the slice interpolation widget is then not managed properly.

This test baselines that entering the ViewControllers module does not modify
which volumes are being displayed in the Slice Views.

A future checkin will test whether the interpolation buttons are accessible.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@21244 3bd1e089-480b-0410-8dfb-8563597acbee