ENH: Speed up VR module initialization
When having a color node with a lot of colors > 1000, deep copying
vtkPiecewiseFunction fires lots of Modified events that are propagated by
the volume property node. This significantly slows down the
initialization of the volume rendering module (20secs in release mode).
The core problem is that a modified volume property node triggers
unnecessary updates. But it will be fixed later on.

Issue #2642 (0006854)

git-svn-id: http://svn.slicer.org/Slicer4/trunk@21278 3bd1e089-480b-0410-8dfb-8563597acbee