BUG: fix crash in tests with scene views (issue #2512)
The scene has a ReferencingNodes vector of bare C pointers to mrml
nodes which were becoming invalid when the nodes were cleared
before the scene.  This change ensures that the pointers in the
ReferencingNodes vector are always registered, so that the nodes
are not freed until we are finished using the pointers.