BUG: fix leaks on scene close - related to issue #2512
Manage the reference counts on nodes that are in the
ReferencingNodes vector so that they don't leak or
get double-freed as a result of operations on the scene.

Also STYLE: add at least some documentation to the header
file describing what these methods are meant to do and
how the data structures are managed.