BUG: fix fiducial display in light box mode.
Use the renderer from the light box proxy manager if it's defined
When in light box mode, switch to using 2d point handles for fiducials.
Avoid seed widget and representation having different renderers on creation.
Unify light box index calculation.
Fix the annotations showing up in the 0th light box when the annotation
should be in the -1th. 
Take into account volume spacing, calculate the distance between the fiducial
and the slice at this light box index using distance from the fiducial point and the slice plane.
Reset the glyph type when changing in and out of lightbox mode.

Issue #1690


git-svn-id: http://svn.slicer.org/Slicer4/trunk@21494 3bd1e089-480b-0410-8dfb-8563597acbee