BUG: make sure precache tags are set for database instances
Since the precache tags are specific to the ctkDICOMDatabase,
be sure to set the correct values to both slicer.app.dicomDatabase
and the database that is internal to the ctkDICOMAppWidget.

Need to use a timer here, because we want to ensure that all plugins
(which are modules) have been discovered before we set the precache
tags.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@21526 3bd1e089-480b-0410-8dfb-8563597acbee