COMP: Set ITKv4 as default version. Update SimpleITK version to v0.6.1
Thanks to the work of the ITK, Slicer and DCMTK community members :),
the default version of ITK built wihtin Slicer will be:

   ITK 4.3 based of commit c31337 (01-06-2013)


Remarks
=======

- While this is a major change from the developer perspective, from
the user perspective, we expect the functionality to behave similarly
or better.  That is why we maintained the major version number to "4".


- In your EXISTING build tree, consider:

      a) manually updating the CMake option ITK_VERSION_MAJOR from 3 to 4

      b) deleting ITKv3* directories. That would help minimizing future
         confusion.


- Extensions superbuild-ing ITKv4 SHOULD be updated.


- It is known that some test are failing exclusively with ITKv4,
these ones will be addressed by the respective module/code maintainers.
See [1] for more details.


- Test have been done to ensure the extension build system was behaving
as expected when using ITKv4. That said, it is possible we experience
some glitch. Make sure to report any issue so that we can solve them
as soon as possible.


- Issue building extension depending on DCMTK has been identified.
For example, extension "houghTransformCLI" failed to build. This will
be addressed as soon as possible by working with ITKv4 and DCMTK
community.


- Enabling SimpleITK is *NOT* enabled by default and has *NOT* been
tested. That said, if you enable the option Slicer_USE_SimpleITK,
the latest version of SimpleITK will be built.


- On windows, if you encounter strange link/build error. Consider
running the following script. We are in the process of identifying
what would be the optimum build and source folder length and your
help and comments are welcome.
See https://github.com/jcfr/list-longest-paths#readme


[1] https://docs.google.com/a/kitware.com/spreadsheet/ccc?key=0AqcFo-C6Ob4-dFJRdWw3LVN6R1VYQlotNXRVaGVVT1E#gid=0

git-svn-id: http://svn.slicer.org/Slicer4/trunk@21688 3bd1e089-480b-0410-8dfb-8563597acbee