ENH: WIP: try moving module loading out of critical path
Need to set up moduleFactoryManager in SlicerAppMain and send
it to the slot.

Then in the slot need to break out the operations into chunks
that are integrated with the event loop.  Also figure out if
some operations can be done in threads.  Certainly CLI
discovery can be broken into multiple steps rather than blocking
the UI while waiting for the process to load.

Need to make sure Volumes Module is loaded so the IO will work as soon
as the UI is loaded.  Want to be able to load data from the command
line before all the modules have loaded.