BUG: try to fix test failures on some platforms
This test was failing on the windows factory machines but not
on non-windows.  Possibly this is due to the window size or another related
variable causing the painting to happen outside the visible
region of the MR in the slice view.  This change makes sure the clicks
occur near the center of the slice view so as to avoid the
boundary case.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@22080 3bd1e089-480b-0410-8dfb-8563597acbee