BUG: Ensure qSlicerStyle is used when using SlicerApp-Real
When the launcher isn't used, for example when using SlicerApp-Real
directly, the plugin library cannot be found before the qSlicerApplication
is created. (QT_PLUGIN_PATH isn't set).

However once the application is started, the current location of the
executable is automaticaly added to the application library paths by Qt.

See Qt-doc: http://qt-project.org/doc/qt-4.7/qcoreapplication.html#id-a4ff2599-e40c-4218-82e7-14e570ebf822
"The directory of the application executable (NOT the working directory) is always added (...)"

Since this also corresponds to the location that needs to be set for the
plugins, the styles are automaticaly discovered.

See also the discussion:
http://massmail.spl.harvard.edu/public-archives/slicer-devel/2013/013320.html

Fixes #3252

git-svn-id: http://svn.slicer.org/Slicer4/trunk@22240 3bd1e089-480b-0410-8dfb-8563597acbee