BUG: Fix "volume not marked as modified on save after harden transform".
Also, fix "annotations not marked as modified on save after they are changed" and "volume not marked as modified on save after volume origin or spacing is changed". Anytime a vtkMRMLStorableNode property that is saved in file is modified, the StorableModifiedTime time stamp must be modified.
Fixes #2980
Fixes #3246
Fixes #3260

From: Yuzheng Zhou <yuzheng.zhou@kitware.com>

git-svn-id: http://svn.slicer.org/Slicer4/trunk@22297 3bd1e089-480b-0410-8dfb-8563597acbee