ENH: Deprecating annotation fiducials
Edit properties in the Data module will go to the Markups module for
both Markups and Annotations fiducials, so that annotation fids
can be converted.

Unregister the Annotations fiducials IO, Markups can read .acsv files.
Add a label in the Annotations GUI that fiducials can be found in the Markups module.

See issue #2606

git-svn-id: http://svn.slicer.org/Slicer4/trunk@22320 3bd1e089-480b-0410-8dfb-8563597acbee