BUG: Fix "No handleSslErrors" slot run time error
Since by default moc (MetaObject compiler) doesn't understand ifdef, let
simply expect that SSL headers are available and unconditionally include
the "handleSslErrors". For this to work, we would have to associated a
compile definition with moc.

Since the QSslError header was already unconditionally included, defining
the slot is not a problem.

Error was:
 Object::connect: No such slot qSlicerWebWidget::handleSslErrors(QNetworkReply*, const QList<QSslError> & ) in /home/jchris/Projects/Slicer/Base/QTGUI/qSlicerWebWidget.cxx:99
 Object::connect:  (receiver name: 'qSlicerWebWidget')

From: Jean-Christophe Fillion-Robin <jchris.fillionr@kitware.com>

git-svn-id: http://svn.slicer.org/Slicer4/trunk@22782 3bd1e089-480b-0410-8dfb-8563597acbee