ENH: adjust the scale slider widgets to volume slice spacing
On entry to the Markups module, check for a background volume
in the red slice viewer, get the slice spacing and adjust
the scale slider maximum values by a scale factor times
the maximum slice spacing.
Round the max scale to the nearest 10 when resetting it from volume
spacing.
Also ensure that if the module is entered with scale values
on the display node that are larger than the slider maximums
that the sliders are adjusted.
Also works for the default settings being larger than slider
maximums.
Set the scale factor to 10 to allow for large markups on
widely spaced volumes.

Issue #3547



git-svn-id: http://svn.slicer.org/Slicer4/trunk@22919 3bd1e089-480b-0410-8dfb-8563597acbee