ENH: Add publishExtension to SlicerWizard
Add a new 'command' to SlicerWizard to publish (to github) an extension.
This creates a git repository for the extension (if one does not already
exist), creates a github repository for the extension, changes the
extension meta-information to reference the github repository, commits
said changes (or makes an initial commit with the contents of the
extension tree), adds the github repository as a remote, and pushes to
github.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@22957 3bd1e089-480b-0410-8dfb-8563597acbee