ENH: Refactor and improve SlicerWizard arg parsing
Move conversion of Python method arguments to CLI arguments to its own
helper function in the Utilities module. Tweak this to handle non-string
arguments and to better handle bool and None arguments (note that a
named argument with the value None is now ignored, as is a named
argument with a value False; True is now used for flag options). Also
add support for "short" arguments (assuming a one-character name is
"short"). All this should make the conversion more intuitive and easier
to use.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@22992 3bd1e089-480b-0410-8dfb-8563597acbee