BUG: #3561 handle paths for windows
Earlier versions of the tpycl (tcl/python) layer worked on all
platforms, but apparently a change occured (maybe the
cmakeified python?) that led to a different interpretation
of this path and now native windows paths with backslashes
are returned from os.path.  As a result we need to convert the
paths before sending them to tcl, which always uses forward
slashes for standard directory operations on all platforms.