WIP: adds python scripted vtk actor
Basically works by allowing you to run arbitrary
python code in the middle of the vtk rendering
pipeline.

Based on the corresponding code from Slicer3

https://www.slicer.org/slicerWiki/index.php/Slicer3:Python:ScriptedActor
http://viewvc.slicer.org/viewvc.cgi/Slicer3/trunk/Modules/ScriptedModule/

Does not work on mac to call PyOpenGL, due to lack of ctypes
in cmakeified python on mac.  Upgrading python to get ctypes
working is stalled on an issue with python setup tools.

https://github.com/davidsansome/python-cmake-buildsystem/issues/18#issuecomment-108648972