ENH: add utility methods to get at code and scheme, clean up
Added utillity methods to allow calls on teh color logic from python to access
the codes and scheme designator strings in the terminology.
Cleaned up the accessors, return an empty string instead of N/A if not found,
put the default color node name in the header file. Don't show an empty
terminology tool tip in the Colors module.
Added testing the utility methods.

Issue #4047